packages/ember-core/pyproject.toml
[project]
name = "ember-core"
dynamic = ["version"]
description = "Core contracts and optional utilities for modular Gaussian splatting."
readme = "README.md"
requires-python = ">=3.11"
license = "Apache-2.0"
license-files = ["LICENSE"]
authors = [
    { name = "Paul Schlack" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "beartype>=0.22.9",
    "jaxtyping>=0.3.9",
    "numpy>=2.3.5",
    "plyfile>=1.1.3",
    "pydantic>=2.11.0",
    "pyyaml>=6.0.3",
    "torch>=2.11.0",
    "tqdm>=4.67.0",
]

[project.urls]
Repository = "https://github.com/panicPaul/ember"
Issues = "https://github.com/panicPaul/ember/issues"

[project.optional-dependencies]
data = [
    "huggingface_hub>=0.34.0",
    "opencv-python-headless>=4.10.0",
    "Pillow>=11.0.0",
]
training = []
eval = []
viewer = [
    "marimo-3dv>=0.1.3",
]
all = [
    "huggingface_hub>=0.34.0",
    "marimo-3dv>=0.1.3",
    "opencv-python-headless>=4.10.0",
    "Pillow>=11.0.0",
]
cu128 = [
    "torch>=2.11.0",
]
cu130 = [
    "torch>=2.11.0",
]

[build-system]
requires = ["hatchling>=1.27.0", "hatch-vcs>=0.5.0"]
build-backend = "hatchling.build"

[tool.uv]
conflicts = [
    [
        { extra = "cu128" },
        { extra = "cu130" },
    ],
]

[tool.uv.sources]
marimo-3dv = { path = "../../packages/marimo-3dv", editable = true }
torch = [
    { index = "pytorch-cu128", extra = "cu128" },
    { index = "pytorch-cu130", extra = "cu130" },
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.hatch.version]
source = "vcs"
fallback-version = "0.0.1a0"

[tool.hatch.version.raw-options]
root = "../.."
version_scheme = "no-guess-dev"

[tool.hatch.build.hooks.vcs]
version-file = "src/ember_core/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/ember_core"]
include = ["src/ember_core/assets/**"]

[tool.hatch.build.targets.sdist]
include = ["src/ember_core/assets/**", "src/ember_core/**/*.py"]
